fix bug 36089, take 2

+++ includes/SMW_CompatibilityHelpers.php
@@ -92,9 +92,9 @@ class SMWCompatibilityHelpers {
 					$pagedbkey = str_replace( ' ', '_', $dbkeys[0] );
 					return new SMWDIWikiPage( $pagedbkey, SF_NS_FORM, '' );
 				} elseif ( count( $dbkeys ) >= 5 ) { // with subobject name (and sortkey)
-					return new SMWDIWikiPage( $dbkeys[0], floatval( $dbkeys[1] ), $dbkeys[2], $dbkeys[4] );
+					return new SMWDIWikiPage( $dbkeys[0], intval( $dbkeys[1] ), $dbkeys[2], $dbkeys[4] );
 				} elseif ( count( $dbkeys ) >= 3 ) { // without subobject name (just for b/c)
-					return new SMWDIWikiPage( $dbkeys[0], floatval( $dbkeys[1] ), $dbkeys[2] );
+					return new SMWDIWikiPage( $dbkeys[0], intval( $dbkeys[1] ), $dbkeys[2] );
 				}
 				break;
 			case SMWDataItem::TYPE_CONCEPT:
